requirements.txt
streamlit
pandas
plotly
fredapi
python-dotenv